+++ packages/sdk-common/src/utils/blockchain.ts
@@ -2,6 +2,7 @@ import { Blockchain } from "@rarible/api-client/build/models/Blockchain"
 
 export type EVMBlockchain = Blockchain.ETHEREUM | Blockchain.POLYGON | Blockchain.MANTLE
 | Blockchain.ARBITRUM | Blockchain.ZKSYNC | Blockchain.CHILIZ | Blockchain.LIGHTLINK
+| Blockchain.RARI
 export const EVMBlockchains: EVMBlockchain[] = [
 	Blockchain.ETHEREUM,
 	Blockchain.POLYGON,
@@ -10,6 +11,7 @@ export const EVMBlockchains: EVMBlockchain[] = [
 	Blockchain.ZKSYNC,
 	Blockchain.CHILIZ,
 	Blockchain.LIGHTLINK,
+	Blockchain.RARI,
 ]
 
 /**