+++ packages/sdk/src/config/development.ts
@@ -13,4 +13,5 @@ export const developmentConfig: RaribleSdkConfig = {
 	zksync: "testnet-zksync",
 	chiliz: "testnet-chiliz",
 	lightlink: "testnet-lightlink",
+	rari: "testnet-rari",
 }