+++ packages/ethereum/test-common/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.67](https://github.com/rarible/sdk/compare/v0.13.66...v0.13.67) (2024-01-10)
+
+**Note:** Version bump only for package @rarible/ethereum-sdk-test-common
+
+
+
+
+
+## [0.13.66](https://github.com/rarible/sdk/compare/v0.13.65...v0.13.66) (2024-01-09)
+
+
+### Features
+
+* **PLTF-216:** support rari blockchain ([#586](https://github.com/rarible/sdk/issues/586)) ([deeb2bc](https://github.com/rarible/sdk/commit/deeb2bc05dbca36500a26e0a7bd58843ab231d25))
+
+
+
+
+
 ## [0.13.65](https://github.com/rarible/sdk/compare/v0.13.64-beta.3...v0.13.65) (2023-12-20)